requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
psycopg2-binary==2.9.9
bcrypt==4.1.3
PyJWT==2.8.0
